utilities/fastq_splitter/requirements.txt
google-cloud-storage==1.13.0